requirements/base.txt
django==1.11.17
psycopg2-binary==2.7.6.1
bpython==0.17.1
django-braces==1.13.0
django-model-utils==3.1.2
django-pipeline==1.6.14
django-widget-tweaks==1.4.3
Unipath==1.1
django-crispy-forms==1.6.1
djangorestframework==3.9.0
django-extensions==1.7.9 # needed for UUID field
celery==4.2.1
django-annoying==0.10.4
requests==2.20.1
django-zipview==1.2.0
elasticsearch>=2.0.0,<3.0.0
elasticsearch-dsl>=2.0.0,<3.0.0
pylibmc==1.6.0
python-dateutil==2.6.0 # pinned for elasticsearch-dsl
raven==6.0.0
reportlab==3.4.0
openpyxl==2.4.7
